+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce incremental backup strategy for raw backups
     with new options `--incremental` and `--incremental-backup`.
 - Add option `--skip-root-check` for create/download/restore (raw) backups.
+- Add option `--show-progress` for file transfers (uses `--info=progress2` for rsync operations).
+- Add/use option `-V, --verbose` for rsync operations.
 
 ### Changed
 - Verbose option changed from `-a, --all` to `-V, --verbose`.